requirements.txt
torch==2.0.1
diffusers==0.24.0
transformers==4.33.0
accelerate==0.23.0
fastapi==0.103.1
uvicorn==0.23.2
pillow==10.0.0
python-multipart==0.0.6
